refactor: makerWallet 수정
